Feature: Add new parameter emr_backoff.
This parameter is used instead of `check_cluster_every` during EMR
throttling retries.

Previously, when throttled, any EMR api call would take
check_cluster_every*1.5^{retry count}. For the default of 30 and, for example,
after 10 retries the next attempt would be in approximiately 30 minutes. This
backoff is too long for certain use-cases. As such we add functionality to be
able to adjust the throttling initial backoff. Therefore, if this is 5 after 10
retries the next attempt would be in only 5 minutes. In the current code we can
only adjust `check_cluster_every`. This is insufficient as this also affects
termination/step completion polling. Therefore we add `emr_backoff`.