Bug fix: Recursively merge extra_cluster_params dict.
The function `_cluster_kwargs` builds a dict of options to submit to AWS
for cluster creation. The last operation the function performs is a dict update
with any custom parameters the user adds. Since this update is not recursive,
if the user sets a nested option that mrjob specifies, then user will override
everything in that nested option. Therefore we must perform a recursive
update. For example, the following will only end up setting the master security
group and not the ec2 key pair nor subnet:

```
    ec2_key_pair: key_name
    subnet: subnet-foo

    extra_cluster_params:
      Instances.EmrManagedMasterSecurityGroup: sg-bar
```